V0.7
Turret dmg and projectiles

+++ Assets/Scripts/Minion2.cs
@@ -7,7 +7,7 @@ public class Minion2 : Enemy
     bool isDamaged = false;
     Color damagedColor = new Color(255f, 0f, 0f, 0.6f);
 
-    public void SetValues(int hp, int damage, int value)
+    public void SetValues(float hp, float damage, float value)
     {
         this.FullHP = hp;
         this.DMG = damage;
